Runtime: separate utility host functions and tests into separate mod.
