+++ crates/runtime/src/logic/host_functions.rs
@@ -1,10 +1,9 @@
 
 mod blobs;
+mod governance;
 mod storage;
 mod system;
-mod governance;
+mod utility;
 
 pub use blobs::*;
-pub use storage::*;
 pub use system::*;
-pub use governance::*;
